+++ src/agents_and_teams/team.cpp
@@ -5,7 +5,7 @@ Team::Team(ParameterSet x){
     p = x;
     // Make a vector of the appropriate length for storing things.
     best_solution.assign(p.max_iter/p.n_agents,
-                         std::vector<long double>(static_cast <unsigned long> (Solution::number_of_objectives, 0.0)));
+                         std::vector<long double>(Solution::number_of_objectives, 0.0));
 }
 
 //// Give the team a new start
@@ -23,7 +23,7 @@ void Team::new_start(void){
 
     // Instantiate the sharing vectors for agents
     Agent::all_current_solutions.assign(p.n_agents, Solution());
-    Agent::all_current_objective_weightings.assign(p.n_agents, std::vector<long double>(static_cast <unsigned long> (Solution::number_of_objectives), 0.0));
+    Agent::all_current_objective_weightings.assign(p.n_agents, std::vector<long double>(Solution::number_of_objectives, 0.0));
 
     // Give agents starting locations
     for(int i=0; i<p.n_agents; i++){
